requirements.txt
# Core libraries
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
pandas==1.5.3
numpy==1.24.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6

# Data processing and visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.18.0

# Forecasting models
prophet==1.1.4
scikit-learn==1.3.0
joblib==1.3.2
shap==0.42.1

# Database connection
supabase==1.0.3
httpx==0.23.3  # Compatible with Supabase 1.0.3

# Testing and development
pytest==7.4.0
coverage==7.3.2

# Miscellaneous
tqdm==4.66.1
